requirements.txt
accelerate==0.28.0
alabaster==0.7.16
appdirs==1.4.4
archspec @ file:///croot/archspec_1709217642129/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
Babel==2.14.0
blinker==1.7.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///work/perseverance-python-buildout/croot/brotli-split_1698805593785/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///work/perseverance-python-buildout/croot/click_1698845879718/work
cloudpickle==0.4.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
conda @ file:///croot/conda_1708369113911/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1698882886606/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///work/perseverance-python-buildout/croot/conda-package-handling_1698851267218/work
conda_package_streaming @ file:///work/perseverance-python-buildout/croot/conda-package-streaming_1698847176583/work
contourpy==1.2.0
cryptography @ file:///croot/cryptography_1710350347627/work
cycler==0.12.1
debugpy @ file:///work/perseverance-python-buildout/croot/debugpy_1698884710808/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
diffusers==0.25.1
distro @ file:///work/perseverance-python-buildout/croot/distro_1701732366176/work
docker-pycreds @ file:///Users/ktietz/demo/mc3/conda-bld/docker-pycreds_1630654474270/work
docutils==0.17.1
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
ExifRead==2.1.2
filelock==3.13.1
Flask==2.3.2
fonttools==4.49.0
fpdf2==2.4.6
fsspec==2024.2.0
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///work/perseverance-python-buildout/croot/gitpython_1698887140758/work
huggingface-hub==0.21.4
idna @ file:///work/perseverance-python-buildout/croot/idna_1698845632828/work
imagesize==1.4.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1709821103657/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1708996548741/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1709559745751/work
ipywidgets @ file:///croot/ipywidgets_1709574692113/work
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
joblib==0.14.1
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257406420/work
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1709322880313/work
kiwisolver==1.4.5
libmambapy @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1709548307542/work/libmambapy
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline @ file:///work/perseverance-python-buildout/croot/matplotlib-inline_1698864771271/work
menuinst @ file:///croot/menuinst_1706732933928/work
mpmath==1.3.0
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==2.5
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opencv-python==4.9.0.80
packaging @ file:///work/perseverance-python-buildout/croot/packaging_1698805390494/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools @ file:///Users/ktietz/demo/mc3/conda-bld/pathtools_1629713893697/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow==10.2.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==3.20.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1705722396628/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py==1.11.0
pycosat @ file:///work/perseverance-python-buildout/croot/pycosat_1698863456259/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyparsing==3.1.2
pyproj==3.6.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytest==3.0.7
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
regex==2023.12.25
requests @ file:///croot/requests_1707355572290/work
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
safetensors==0.4.2
scipy==1.12.0
sentry-sdk @ file:///work/perseverance-python-buildout/croot/sentry-sdk_1698901418049/work
setproctitle @ file:///work/perseverance-python-buildout/croot/setproctitle_1698901493537/work
setuptools==68.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
snowballstemmer==2.2.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
tokenizers==0.15.2
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchinfo==1.8.0
torchvision==0.17.1+cu118
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1708363096407/work
tqdm @ file:///work/perseverance-python-buildout/croot/tqdm_1701735729845/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1710254411456/work
transformers==4.38.2
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
urllib3 @ file:///croot/urllib3_1707770551213/work
wandb==0.16.4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.1
wheel==0.41.2
widgetsnbextension @ file:///croot/widgetsnbextension_1709322880396/work
xformers==0.0.24
xmltodict==0.10.2
zipp @ file:///work/perseverance-python-buildout/croot/zipp_1707367835224/work
zstandard @ file:///work/perseverance-python-buildout/croot/zstandard_1698847073368/work
